Rename whitespectre.md to whitespectre
We're Whitespectre. We partner with companies to develop high-impact products and platforms. We follow agile methodologies and specialize in Ruby on Rails and Mobile applications.

We combine technical excellence, a deep understanding of our clients’ business goals, and agile development practices to deliver high-impact products.

Our engagement models vary; from taking end-to-end responsibility for the successful delivery and ongoing iteration of a defined product, to embedding ourselves within client teams to increase their throughput with our best practices and mentoring.
Mission: Define, Deliver, Iterate.
We believe the best products start with clear business drivers and then evolve iteratively, based on data-driven decision making, customer insights, and user-centric design.

Technology Excellence
Our motto: never overengineer, but always design for the future. Scaling technology is an art and a science. We believe in understanding a system's performance requirements and delivering the most simple, robust and maintainable solution to exceed these.